Fixes
by: MetalloloM-Edition
Updated russian translation

by Alcater:
Added Fluid Template tooltip info
Added sulfur to UF6 production outputs
Added Cyclotron plug bonus

Rebalanced Toxic trait

Fixed Jetpack visibility in spectator mode
Fixed RBMK Console 1x1 tooltip
Fixed Gas Cent shiftclicking
Fixed coal and asbestos config
Fixed Hazard stack effect
Fixed Boilers not allowing modded fluids
Fixed filter gui mining drops
Fixed Catalytic Cracker Hitbox
Fixed Miner Rocket lagg
Fixed get fluid color crash
Fixed missing power block recipes
Fixed Gas Cent muffler det
Fixed Chemplant I/O
Fixed Chemplant Fluid rendering
Fixed Skybox render crash
Fixed Decon init crash

+++ src/main/java/com/hbm/explosion/ExplosionLarge.java
@@ -59,7 +59,7 @@ public class ExplosionLarge {
 			fx.motionZ = vec.z;
 			world.spawnEntity(fx);
 			
-			vec = vec.rotateYaw(360 / count);
+			vec = vec.rotateYaw((float) 360 / count);
 		}
 	}
